feat: refactor form to react form hook
