+++ rss-react-app/package.json
@@ -19,6 +19,7 @@
     "gh-pages": "^5.0.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-hook-form": "^7.43.8",
     "react-router-dom": "^6.9.0"
   },
   "devDependencies": {
